+++ R/Result.R
@@ -9,12 +9,19 @@ NULL
 #' @docType methods
 NULL
 
-OdbcResult <- function(connection, statement) {
+OdbcResult <- function(connection, statement, params = NULL, immediate = FALSE) {
   if (nzchar(connection@encoding)) {
     statement <- enc2iconv(statement, connection@encoding)
   }
-  ptr <- new_result(connection@ptr, statement)
-  new("OdbcResult", connection = connection, statement = statement, ptr = ptr)
+  ptr <- new_result(connection@ptr, statement, immediate)
+  res <- new("OdbcResult", connection = connection, statement = statement, ptr = ptr)
+
+  if (!is.null(params)) {
+    on.exit(dbClearResult(res))
+    dbBind(res, params = params)
+    on.exit(NULL)
+  }
+  res
 }
 
 #' @rdname OdbcResult
@@ -49,7 +56,7 @@ setMethod(
 setMethod(
   "dbFetch", "OdbcResult",
   function(res, n = -1, ...) {
-    result_fetch(res@ptr, n, ...)
+    result_fetch(res@ptr, n)
   })
 
 #' @rdname OdbcResult
@@ -88,7 +95,7 @@ setMethod(
 setMethod(
   "dbColumnInfo", "OdbcResult",
   function(res, ...) {
-    result_column_info(res@ptr, ...)
+    result_column_info(res@ptr)
   })
 
 #' @rdname OdbcResult
@@ -111,10 +118,14 @@ setMethod(
 
 #' @rdname OdbcResult
 #' @inheritParams DBI::dbBind
+#' @inheritParams odbc-tables
 #' @export
 setMethod(
   "dbBind", "OdbcResult",
-  function(res, params, ...) {
-    result_bind(res@ptr, as.list(params))
+  function(res, params, ..., batch_rows = getOption("odbc.batch_rows", 1024)) {
+
+    batch_rows <- parse_size(batch_rows)
+
+    result_bind(res@ptr, as.list(params), batch_rows)
     invisible(res)
   })