+++ src/connection.cpp
@@ -49,6 +49,7 @@ Rcpp::DataFrame list_data_sources_() {
 connection_ptr odbc_connect(
     std::string const& connection_string,
     std::string const& timezone = "",
+    std::string const& timezone_out = "",
     std::string const& encoding = "",
     int bigint = 0,
     long timeout = 0) {
@@ -56,6 +57,7 @@ connection_ptr odbc_connect(
       new std::shared_ptr<odbc_connection>(new odbc_connection(
           connection_string,
           timezone,
+          timezone_out,
           encoding,
           static_cast<bigint_map_t>(bigint),
           timeout)));
@@ -64,13 +66,15 @@ connection_ptr odbc_connect(
 std::string get_info_or_empty(connection_ptr const& p, short type) {
   try {
     return (*p)->connection()->get_info<std::string>(type);
-  } catch (nanodbc::database_error c) {
+  } catch (const nanodbc::database_error& c) {
     return "";
   }
 }
 
 // [[Rcpp::export]]
 Rcpp::List connection_info(connection_ptr const& p) {
+  auto getdata_ext =
+      (*p)->connection()->get_info<unsigned short>(SQL_GETDATA_EXTENSIONS);
   return Rcpp::List::create(
       Rcpp::_["dbname"] = get_info_or_empty(p, SQL_DATABASE_NAME),
       Rcpp::_["dbms.name"] = get_info_or_empty(p, SQL_DBMS_NAME),
@@ -84,7 +88,11 @@ Rcpp::List connection_info(connection_ptr const& p) {
       Rcpp::_["odbc.version"] = get_info_or_empty(p, SQL_ODBC_VER),
       Rcpp::_["driver.version"] = get_info_or_empty(p, SQL_DRIVER_VER),
       Rcpp::_["odbcdriver.version"] = get_info_or_empty(p, SQL_DRIVER_ODBC_VER),
-      Rcpp::_["supports.transactions"] = (*p)->supports_transactions());
+      Rcpp::_["supports.transactions"] = (*p)->supports_transactions(),
+      Rcpp::_["getdata.extensions.any_column"] =
+          ((getdata_ext & SQL_GD_ANY_COLUMN) != 0),
+      Rcpp::_["getdata.extensions.any_order"] =
+          ((getdata_ext & SQL_GD_ANY_ORDER) != 0));
 }
 
 // [[Rcpp::export]]