+++ src/odbc_connection.cpp
@@ -21,10 +21,12 @@ void odbc_connection::set_current_result(odbc_result* r) {
 odbc_connection::odbc_connection(
     std::string connection_string,
     std::string timezone,
+    std::string timezone_out,
     std::string encoding,
     bigint_map_t bigint_mapping,
     long timeout)
     : current_result_(nullptr),
+      timezone_out_str_(timezone_out),
       encoding_(encoding),
       bigint_mapping_(bigint_mapping) {
 
@@ -32,9 +34,15 @@ odbc_connection::odbc_connection(
     Rcpp::stop("Error loading time zone (%s)", timezone);
   }
 
+  // timezone_out_ will not be used. This line is just to ensure
+  // the provided value is valid.
+  if (!cctz::load_time_zone(timezone_out, &timezone_out_)) {
+    Rcpp::stop("Error loading timezone_out (%s)", timezone_out);
+  }
+
   try {
     c_ = std::make_shared<nanodbc::connection>(connection_string, timeout);
-  } catch (nanodbc::database_error e) {
+  } catch (const nanodbc::database_error& e) {
     throw Rcpp::exception(e.what(), FALSE);
   }
 }
@@ -72,12 +80,15 @@ bool odbc_connection::is_current_result(odbc_result* result) const {
 bool odbc_connection::supports_transactions() const {
   try {
     return c_->get_info<unsigned short>(SQL_TXN_CAPABLE) != SQL_TC_NONE;
-  } catch (nanodbc::database_error e) {
+  } catch (const nanodbc::database_error& e) {
     return false;
   }
 }
 
 cctz::time_zone odbc_connection::timezone() const { return timezone_; }
+std::string odbc_connection::timezone_out_str() const {
+  return timezone_out_str_;
+}
 std::string odbc_connection::encoding() const { return encoding_; }
 
 bigint_map_t odbc_connection::get_bigint_mapping() const {